Update README.md
Contact manager is a simple console application. It is similar to the contact manager in cell phones.
The key features of contact management system are listed below:
•	Add new contacts: with information such as name, phone number,  and email.
•	List all contacts: lists all the contacts stored in file with their respective contact details.
•	Search contacts: based on name and phone number.
•	Edit contacts: edit information given while adding the contacts – name, phone number and email.
•	Delete contacts: deletes contacts from file.
In addition to the above features of categorize contact to different fields such as family, friends, co-worker and VIP.
VIP contacts are password protected and this category provides user a privacy for some of his required contact.